// src/generated/types/ExamModelDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ExamModelDto = {
  /**
   * @description Exam model ID
   * @type string
   */
  id: string
  /**
   * @description Name of the exam model
   * @type string
   */
  name: string
  /**
   * @description Version of the exam model (A, B, C, D...)
   * @type string
   */
  version: string
  /**
   * @description Description of the exam model
   * @type string | undefined
   */
  description?: string | undefined
  /**
   * @description Number of students assigned to this model
   * @type number
   */
  assignedStudentCount: number
  /**
   * @description Whether the model is active
   * @type boolean
   */
  isActive: boolean
  /**
   * @description Event ID
   * @type string
   */
  eventId: string
  /**
   * @description Array of file information
   * @type array
   */
  files: unknown[]
  /**
   * @description Creation date
   * @type string, date-time
   */
  created_at: Date
  /**
   * @description Last update date
   * @type string, date-time
   */
  updated_at: Date
}